requirements.txt
transformers==5.3.0
accelerate==1.13.0
datasets==2.18.0

openai-whisper

numpy
scipy
pandas
scikit-learn

librosa
soundfile

sentence-transformers
rapidfuzz

streamlit